requirements.txt
annotated-types==0.5.0
anyio==4.0.0
boto3==1.28.40
botocore==1.31.40
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
exceptiongroup==1.1.3
fastapi==0.103.0
h11==0.14.0
idna==3.4
jmespath==1.0.1
pydantic==2.3.0
pydantic_core==2.6.3
python-dateutil==2.8.2
python-dotenv==1.0.0
requests==2.31.0
s3transfer==0.6.2
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.7.1
urllib3==1.26.16
uvicorn==0.23.2
